Rename a file in models directory
